chore: move `MainQueueDispatchDecorator` to separate file
